// src/generated/nf-servico-v1.ts
export type paths = {
// ...
  "/v2/webhooks": {
// ...
    post: {
// ...
      responses: {
        /** @description Sucesso na criação da webhook */
        201: {
          content: {
            readonly "application/json": {
              /** @description WebHook (Notificação HTTP) */
              readonly webHook?: {
                /**
                 * @description ID exclusivo do WebHook. Este ID pode ser usado para se referir mais tarde ao WebHook no caso de
                 * precisa ser atualizado ou excluído. O ID é, por padrão, na forma de um GUID.
                 */
                readonly id?: string;
                /** @description A URL onde as notificações dos eventos deverão entregues. */
                readonly uri: string;
                /**
                 * @description Segredo, contendo de 32 até 64 caracteres que será usado gerar o valor
                 * do **HMAC-SHA1** em hexadecimal que será enviado no cabeçalho HTTP *X-Hub-Signature*.
                 * O HMAC-SHA1 será gerado baseado no bytes do corpo do evento de notificação que será enviado.
                 */
                readonly secret?: string;
                /**
                 * Format: int32
                 * @description WebHook Media Type
                 * @enum {integer}
                 */
                readonly contentType?: 0 | 1;
                /**
                 * @description Determina se o certificado SSL do host da URL será verificado ao entregar as notificações dos eventos.
                 * Defina como **true** para pular a verificação do certificado SSL, sendo o padrão: **false**.
                 */
                readonly insecureSsl?: boolean;
                /**
                 * Format: int32
                 * @description WebHook Status
                 * @enum {integer}
                 */
                readonly status?: 0 | 1;
                /**
                 * @description Lista de filtros sem distinção entre maiúsculas e minúsculas associado a este webhook.
                 * Os filtros são usados para determinar em quais notificações dos eventos esse WebHook será notificado.
                 * Os valores de filtros suportados pode ser consultados através do requisição do **Tipos de Eventos**.
                 */
                readonly filters?: readonly string[];
                /** @description Lista de cabeçalhos HTTP adicionais que serão enviados juntamente com as notificações dos eventos para o webhook. */
                readonly headers?: {
                  [key: string]: string;
                };
                /**
                 * @description Lista de propriedades adicionais que não diferenciam maiúsculas de minúsculas que serão enviadas
                 * juntamente com as notificações dos eventos para o webhook como parte do corpo da entidade de solicitação HTTP.
                 */
                readonly properties?: {
                  [key: string]: unknown;
                };
                /**
                 * Format: date-time
                 * @description Data de criação do webhook
                 */
                readonly createdOn?: string;
                /**
                 * Format: date-time
                 * @description Data de modificação do webhook
                 */
                readonly modifiedOn?: string;
              };
            };
          };
        };
// ...
      };
    };
// ...
  };
// ...
};